requirement.txt
africastalking==1.2.4
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
contextlib2==21.6.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.2
djangorestframework==3.14.0
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
pytz==2022.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
schema==0.7.5
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
